Refactored into ApiCallCounter component to prevent unnecessary re-renders and fetching
